+++ tests/integration_test.rs
@@ -310,6 +310,32 @@ fn uncompress_to_dir_not_preserve_owner() {
 }
 
 #[test]
+fn uncompress_7z_to_dir_not_preserve_owner() {
+    let dir = tempfile::TempDir::new().expect("Failed to create the tmp directory");
+    let mut source = std::fs::File::open("tests/fixtures/tree.7z").unwrap();
+
+    uncompress_archive(&mut source, dir.path(), Ownership::Ignore)
+        .expect("Failed to uncompress the file");
+
+    assert_eq!(
+        dir.path().join("tree/branch1/leaf").exists(),
+        true,
+        "the path doesn't exist"
+    );
+    assert_eq!(
+        dir.path().join("tree/branch2/leaf").exists(),
+        true,
+        "the path doesn't exist"
+    );
+
+    let contents = std::fs::read_to_string(dir.path().join("tree/branch2/leaf")).unwrap();
+    assert_eq!(
+        contents, "Goodbye World\n",
+        "Uncompressed file did not match"
+    );
+}
+
+#[test]
 fn uncompress_to_dir_with_utf8_pathname() {
     let dir = tempfile::TempDir::new().expect("Failed to create the tmp directory");
     let mut source = std::fs::File::open("tests/fixtures/utf8.tar").unwrap();